% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMissingness.R
\name{ComputeMissingness}
\alias{ComputeMissingness}
\title{ComputeMissingness}
\usage{
ComputeMissingness(
  input,
  type = "glyco",
  removeAllNA = TRUE,
  whichPeptide = NA,
  whichAlias = NULL,
  silent = FALSE
)
}
\arguments{
\item{input}{Your formatted GlycoDiveR data.}

\item{type}{Choose between "glyco" or "all"}

\item{removeAllNA}{Remove peptides with only NA values. This can, for example
happen if you upload abundance values using Perseus or MSstats.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{silent}{silence printed information (default = FALSE)}
}
\value{
The percentage of missing data.
}
\description{
This computes the percentage of missing values in your GlycoDiveR uploaded data.
The computation is done on a (glyco)peptide level. Peptides with a quantitative
value of 0 are considered missing.
}
\examples{
\dontrun{ComputeMissingness(mydata, type = "glyco")}
}
