% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterComparison.R
\name{FilterComparison}
\alias{FilterComparison}
\title{FilterComparison}
\usage{
FilterComparison(
  comparison,
  statisticalCutoff = 0.05,
  whichStatistic = "pvalue",
  whichLabel = NA,
  minf = FALSE,
  inf = FALSE
)
}
\arguments{
\item{comparison}{A comparison dataframe.}

\item{statisticalCutoff}{What statistical cutoff to use (default = 0.05)}

\item{whichStatistic}{Typically, "pvalue", or "adjpvalue". Must correspond to
an identically named column in your comparison dataframe.}

\item{whichLabel}{What label(s) to use. Corresponds to the Label column of
the input comparison dataframe.}

\item{minf}{TRUE/FALSE if -inf should be retained (default = FALSE).}

\item{inf}{TRUE/FALSE if inf should be retained (default = FALSE)}
}
\value{
A subset of your comparison dataframe
}
\description{
This functions helps filter your comparison dataframe, which can subsequently
be used for other plots and functions. The dataframe needs to have a column
supplied by the whichStatistic argument (typically, "pvalue", or "adjpvalue").
Alternatively, you can filter the comparison dataframe yourself, for example,
using subset or dplyr::filter.
}
\examples{
\dontrun{FilterComparison(comparison, whichLabel =
c("SampleA-SampleB", "SampleB-SampleC")) }
}
