% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnnotationTemplate.R
\name{GetAnnotationTemplate}
\alias{GetAnnotationTemplate}
\title{GetAnnotationTemplate}
\usage{
GetAnnotationTemplate(path, tool)
}
\arguments{
\item{path}{The path the the folder that holds the subfolder with the PSM files}

\item{tool}{The search engine used. Options: MSFragger, Byonic, pGlyco}
}
\value{
Exported annotation file
}
\description{
This created an annotation.csv file in the path you specify.
}
\examples{
\dontrun{GetAnnotationTemplate("Z:/Subfolder", tool = "MSFragger")}

}
