% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportByonic.R
\name{ImportByonic}
\alias{ImportByonic}
\title{ImportByonic}
\usage{
ImportByonic(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff,
  glycanScoreCutoff,
  deltaModCutoff = 1,
  scrape = TRUE,
  removeReverse = TRUE,
  cutoffFilter = TRUE,
  dropNoQuant = FALSE,
  minPeptideCoverage = FALSE,
  thresholdMode = c("group", "total"),
  GlyToucan = TRUE
)
}
\arguments{
\item{path}{The folder path, which will be used to find all "Byonic.xsls" files
in this folder and all subfolders.}

\item{annotation}{The annotation dataframe. Please generate a template using
GetAnnotationTemplate("path", tool = "Byonic")}

\item{fastaPath}{The path to the fasta file that was used.}

\item{peptideScoreCutoff}{The score cutoff from the "Score" column. The cutoff
is the lower limit.}

\item{glycanScoreCutoff}{The score cutoff in the Log Prob column. The score is the
upper limit.}

\item{deltaModCutoff}{The score cutoff for the Delta Mod score. The score is the
lower limit.}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{removeReverse}{Remove reverse hits as noted with the ">Reverse " tag.}

\item{cutoffFilter}{Filter for the glycan score, peptide score, N Sequon, and
confidence levels.}

\item{dropNoQuant}{Remove or keep PSMs without a quantitative value or have a quant
of 0.}

\item{minPeptideCoverage}{Works together with the thresholdMode argument.
If thresholdMode = "total", provide a single numeric value. Values between 0
and 1 are interpreted as proportions (e.g., 0.1 = 10\%), while values >= 1 are
interpreted as absolute counts (e.g., 2 = at least two peptides identified).
If thresholdMode = "group", provide a numeric vector of length two (e.g., c(0.1, 3)).
The first value is interpreted in the same way as for "total" (proportion or
absolute count). The second value specifies the minimum number of groups in
which this threshold must be met.}

\item{thresholdMode}{Character string specifying the filtering mode: "total"
applies a global threshold across all data, while "group" identifies the threshold
within groups.}

\item{GlyToucan}{set to TRUE will connect to the GlyCosmos API to retrieve
GlyToucan identifiers.}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
This is the importer for Byonic data. It needs to untouched Byonic output, which
is a Excel file with at least a Summary and a Spectra tab. This script will look
for all files in the folder and all subfolders using the name tag "Byonic.xslx".
}
\examples{
\dontrun{ImportByonic(path = "C:/Byonic",
annotation = "C:/annotation.csv",
fastaPath = "C:/fastafile.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 1)}
}
