% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportGlycanFinder.R
\name{ImportGlycanFinder}
\alias{ImportGlycanFinder}
\title{ImportGlycanFinder}
\usage{
ImportGlycanFinder(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff = 0,
  glycanScoreCutoff = 0,
  scrape = TRUE,
  normalization = "median",
  cutoffFilter = TRUE,
  dropNoQuant = FALSE,
  minPeptideCoverage = FALSE,
  thresholdMode = c("group", "total"),
  useExtendedOGlycanCategories = FALSE,
  GlyToucan = TRUE,
  SScoreCutoff = FALSE,
  structureConfidenceCutoff = FALSE,
  AScoreCutoff = FALSE
)
}
\arguments{
\item{path}{Path to search engine output folder}

\item{annotation}{Path to annotation file}

\item{fastaPath}{Path to FASTA file}

\item{peptideScoreCutoff}{Peptide score cutoff, which is the Glycopeptide Score
column in the search engine output file.}

\item{glycanScoreCutoff}{Glycan score cutoff, which is the Glycan Score
column in the search engine output file.}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{normalization}{The (glyco)peptide normalization used.
Choose between "median" (default) or "none".}

\item{cutoffFilter}{Filter for the glycan score, peptide score, and
confidence levels during data import}

\item{dropNoQuant}{Remove or keep PSMs without a quantitative value or have a quant
of 0.}

\item{minPeptideCoverage}{Works together with the thresholdMode argument.
If thresholdMode = "total", provide a single numeric value. Values between 0
and 1 are interpreted as proportions (e.g., 0.1 = 10\%), while values >= 1 are
interpreted as absolute counts (e.g., 2 = at least two peptides identified).
If thresholdMode = "group", provide a numeric vector of length two (e.g., c(0.1, 3)).
The first value is interpreted in the same way as for "total" (proportion or
absolute count). The second value specifies the minimum number of groups in
which this threshold must be met.}

\item{thresholdMode}{Character string specifying the filtering mode: "total"
applies a global threshold across all data, while "group" identifies the threshold
within groups.}

\item{useExtendedOGlycanCategories}{set to FALSE will classify all O-glycans as OGlycan
Set to TRUE will classify them in "Sialofucosylated", "Sialylated", "Fucosylated",
and "OGlycan".}

\item{GlyToucan}{set to TRUE will connect to the GlyCosmos API to retrieve
GlyToucan identifiers.}

\item{SScoreCutoff}{The lower limit of the S score used to filter the data, e.g.
SScoreCutoff = 10. This will retain S Scores that are 10 or higher.}

\item{structureConfidenceCutoff}{The Structure Cutoffs to be retained. For example,
structureConfidenceCutoff = c("LOW", "HIGH")}

\item{AScoreCutoff}{The lower limit of the A score used to filter the data, e.g.
AScoreCutoff = 10. This will retain A Scores 10 or higher. The highest score is
retained for For multiply glycosylated peptides.}
}
\value{
A formatted GlycoDiveR file.
}
\description{
This imports search results output by GlycanFinder. It will work for
N-glycopeptide searches and O-glycopeptides searches. During import, this
function will look for all the glycan.glycopsms.csv files in the
specified folder and all the subfolders.
}
\examples{
\dontrun{
ImportGlycanFinder(
  path = "Z:/Analysis 2",
  annotation = "Z:/annotation.csv",
  fastaPath = "Z:/glycan.proteins.fasta",
  normalization = "none",
  GlyToucan = FALSE,
  structureConfidenceCutoff = "HIGH",
  AScoreCutoff = 10
)
}
}
