% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMSstatsAbundances.R
\name{ImportMSstatsAbundances}
\alias{ImportMSstatsAbundances}
\title{ImportMSstatsAbundances}
\usage{
ImportMSstatsAbundances(input, MSstatsPath, cleanCCarbamidomethylation = TRUE)
}
\arguments{
\item{input}{This is your formatted GlycoDiveR data}

\item{MSstatsPath}{The filepath to the MSstats generated csv.}

\item{cleanCCarbamidomethylation}{Removes \[57.0215\] from the MSstats
ModifiedPeptide column (default = TRUE)}
}
\value{
your input data with updated quantitative values.
}
\description{
This script uses MSstats generated quantitative data and uses this data to fill
in the existing quantitative values. It will only retain quantitative values
present in the MSstats data. Requirements are 1) matching Alias column with the
same values as in your GlycoDiveR data, 2) MSstats data has a LogIntensities
column, 3) MSstats data has a ModifiedPeptide column with values correspoding
to the ModifiedPeptide column in your data. This script is for MSstats data,
but should work for other inputs too.
}
\examples{
\dontrun{ImportMSstatsComparison(mydata, "C:/User/Pathtocsv.csv")}
}
