% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPerseusComparison.R
\name{ImportPerseusComparison}
\alias{ImportPerseusComparison}
\title{ImportPerseusComparison}
\usage{
ImportPerseusComparison(input, path, cleanCCarbamidomethylation = TRUE)
}
\arguments{
\item{input}{Your already imported GlycoDiveR data.}

\item{path}{The path to the Perseus .txt file.}

\item{cleanCCarbamidomethylation}{Removes \[57.0215\] from the MSstats
ModifiedPeptide column (default = TRUE)}
}
\value{
A comparison dataframe.
}
\description{
This imports a Perseus comparison and returns it as a dataframe. The Conditions
specified in Perseus must match the Condition values in the GlycoDiveR annotation
file. Perseus formats column names as such: Condition1_Condition2. GlycoDiveR
creates identical mapping.
}
\examples{
\dontrun{ImportPerseusComparison(mydata, C:/PerseusOutput.txt)}
}
