% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportpGlyco.R
\name{ImportpGlyco}
\alias{ImportpGlyco}
\title{ImportpGlyco}
\usage{
ImportpGlyco(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff = 0.01,
  glycanScoreCutoff = 0.01,
  normalization = "median",
  scrape = TRUE,
  cutoffFilter = TRUE
)
}
\arguments{
\item{path}{The folder path, which will be used to find all
'pGlycoDB-GP-FDR-Pro-Quant-Site.txt' files in this folder and all subfolders.}

\item{annotation}{The annotation dataframe. Please generate a template using
GetAnnotationTemplate("path", tool = "pGlyco")}

\item{fastaPath}{The path to the fasta file that was used.}

\item{peptideScoreCutoff}{The score cutoff for the "PeptideFDR" column. The cutoff
is the upper limit.}

\item{glycanScoreCutoff}{The score cutoff for the "GlycanFDR" column. The score is the
upper limit.}

\item{normalization}{The (glyco)peptide normalization used.
Choose between "none" or "median" (default = "median").}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{cutoffFilter}{Filter for the glycan score, peptide score, N Sequon, and
confidence levels.}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
This is the importer for pGlyco data. It needs to untouched pGlyco output, which
is a .txt file named 'pGlycoDB-GP-FDR-Pro-Quant-Site.txt'. This script will look
for all files in the folder and all subfolders.
}
\examples{
\dontrun{ImportpGlyco(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0.01,
glycanScoreCutoff = 0.01,
scrape = FALSE)}
}
