% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageColorScheme.R
\name{ManageColorScheme}
\alias{ManageColorScheme}
\title{ManageColorScheme}
\usage{
ManageColorScheme(
  edit = NULL,
  action = c("View", "Append", "Replace", "Reset")
)
}
\arguments{
\item{edit}{Supply a vector containing R supported colors. These can for example
be "white" or "#E1D89F".}

\item{action}{Choose between the following:
View: to view the current modification database.
Append: append to the currrent modification database with the supplied edit.
Replace: replace the modification database with the supplied edit.
Reset: reset to the default values.}
}
\value{
The updated database.
}
\description{
Manage the color palette used for plotting
}
\examples{
\dontrun{
ManageColorScheme(edit = c("white", "black", "#E1D89F"), action="Append")
}
}
