% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageGlycanCategories.R
\name{ManageGlycanCategories}
\alias{ManageGlycanCategories}
\title{Manage glycan categories}
\usage{
ManageGlycanCategories(
  edit = NULL,
  action = c("View", "Replace", "Reset", "showCode")
)
}
\arguments{
\item{edit}{A data frame used to replace the existing glycan categories.
You can generate a template for this data frame using \code{action = "showCode"}.
The columns for the dataframe are:

\describe{
\item{Priority}{Numeric priority used to resolve matches. The category with the highest priority is assigned.}
\item{Residue}{Amino acid at the glycosylation site.}
\item{Pattern}{Pattern matched against the \code{TotalGlycanComposition} column (see \code{mydata$PSMTable$TotalGlycanComposition}).
The pattern is a Regular Expressions (Regex). AI is super helpful getting the Regex
that you need.}
\item{Min_H / Max_H}{Minimum and maximum number of hexoses.}
\item{Min_N / Max_N}{Minimum and maximum number of HexNAc residues.}
\item{Category}{Assigned glycan category for the highest-priority match.}
}

If you introduce new glycan categories, use \code{ManageGlycanColors()} to assign
colors to them.}

\item{action}{Character string specifying the action to perform. One of:
\describe{
\item{"View"}{Return the current glycan categories.}
\item{"Replace"}{Replace the current glycan categories with \code{edit}.}
\item{"Reset"}{Reset glycan categories to the default values.}
\item{"showCode"}{Return example code for creating a new \code{edit} data frame.}
}}
}
\value{
A data frame containing the (updated) glycan categories.
}
\description{
Manage the glycan categories used by GlycoDiveR. Run \code{ManageGlycanCategories(action = "View")}
to see what this database looks like. To use custom glycan categories, please run this function before
you import your data. The colors for new glycan categories should be registered using \code{ManageGlycanColors()}.
}
\examples{
\dontrun{
# View current categories
ManageGlycanCategories(action = "View")

# Replace categories using a template
ManageGlycanCategories(
  edit = df_from_action_equals_showCode,
  action = "Replace"
)
}
}
