% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageGlycanColors.R
\name{ManageGlycanColors}
\alias{ManageGlycanColors}
\title{ManageGlycanColors}
\usage{
ManageGlycanColors(
  edit = NULL,
  action = c("View", "Append", "Replace", "Reset")
)
}
\arguments{
\item{edit}{Supply a path the a csv file containing the headers "GlycanType",
and "color", or supply a data frame with those headers.}

\item{action}{Choose between the following:
View: to view the current modification database.
Append: append to the currrent modification database with the supplied edit. If
duplicate GlycanType values are found, it will overwrite the existing values.
Replace: replace the modification database with the supplied edit.
Reset: reset to the default values.}
}
\value{
The updated database.
}
\description{
Manage the color palette used for glycans.
}
\examples{
\dontrun{
ManageGlycanColors(edit = mydataframe, action="Append")
}
}
