% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageGlycanDatabase.R
\name{ManageGlycanDatabase}
\alias{ManageGlycanDatabase}
\title{ManageGlycanDatabase}
\usage{
ManageGlycanDatabase(
  edit = NULL,
  action = c("View", "Append", "Replace", "Reset")
)
}
\arguments{
\item{edit}{Supply a path the a csv file containing the headers "FullName",
"ShortName" and "GMass", or supply a data frame with those headers.}

\item{action}{Choose between the following:
View: to view the current modification database.
Append: append to the currrent modification database with the supplied edit.
Replace: replace the modification database with the supplied edit.
Reset: reset to the default values.}
}
\value{
The updated database.
}
\description{
Manage the glycan database.
}
\examples{
\dontrun{
ManageGlycanDatabase(edit = "C:/locationOfDataframe.csv", action="Append")
}
}
