% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageModificationDatabase.R
\name{ManageModificationDatabase}
\alias{ManageModificationDatabase}
\title{ManageModificationDatabase}
\usage{
ManageModificationDatabase(
  edit = NULL,
  action = c("View", "Append", "Replace", "Reset")
)
}
\arguments{
\item{edit}{Supply a path the a csv file containing the headers "FullName" and
"ModificationMass", or supply a data frame with the headers "FullName" and
"ModificationMass".}

\item{action}{Choose between the following:
View: to view the current modification database.
Append: append to the currrent modification database with the supplied edit.
Replace: replace the modification database with the supplied edit.
Reset: reset to the default values.}
}
\value{
The updated database.
}
\description{
Manage the modification database.
}
\examples{
\dontrun{
ManageModificationDatabase(edit = mydataframe, action="Append")
}
}
