% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFraggerImporter.R
\name{MSFraggerImporter}
\alias{MSFraggerImporter}
\title{Title}
\usage{
MSFraggerImporter(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff,
  glycanScoreCutoff,
  scrape = FALSE,
  normalization = "median",
  convertFPModCodeToMass = TRUE,
  filterForNoNSequon = FALSE,
  OPairLevelConversion = c(0, 0, 0.05, 0.1)
)
}
\arguments{
\item{path}{Path to search engine output folder}

\item{annotation}{Path to annotation file}

\item{fastaPath}{Path to FASTA file}

\item{peptideScoreCutoff}{Peptide score cutoff}

\item{glycanScoreCutoff}{Glycan score cutoff}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data}

\item{normalization}{The (glyco)peptide modification.
Choose between "none" or "median"}

\item{convertFPModCodeToMass}{MSFragger uses modification code in peptide
modified sequences. This replaces the code with the mass of the modification.
Keep this enabled when importing MSstats comparison results as MSstats uses
modification masses instead of modification codes. Default = TRUE}

\item{filterForNoNSequon}{Filter for peptides without an N-sequon. Only works
for OPair data (default = FALSE).}

\item{OPairLevelConversion}{Convert the OPair site probability levels to glycan
q-scores. Provide a vector that has values for each level. The default is
c(0,0,0.05,0.1), which means level1 = 0, level1b = 0, level2 = 0.05, and
level3 = 0.1). This qscore filtering is used with the Importers
glycanScoreCutoff argument.}
}
\value{
Formatted dataframes
}
\description{
Title
}
\examples{
\dontrun{MSFraggerImporter(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 0.05,
scrape = FALSE)}
}
