% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompletenessHeatmap.R
\name{PlotCompletenessHeatmap}
\alias{PlotCompletenessHeatmap}
\title{Plot Completeness Heatmap}
\usage{
PlotCompletenessHeatmap(input, grouping = "peptide", peptideType = "glyco")
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{peptide or glyco}

\item{peptideType}{glyco or other}
}
\value{
Grouped heatmap
}
\description{
Plot Completeness Heatmap
}
\examples{
PlotCompletenessHeatmap(mydata)
}
