% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotErrorVersusRt.R
\name{PlotErrorVersusRt}
\alias{PlotErrorVersusRt}
\title{PlotErrorVersusRt}
\usage{
PlotErrorVersusRt(
  input,
  type = c("both", "glyco", "nonGlyco"),
  upperLimit = NA,
  lowerLimit = NA,
  alpha = 0.5,
  whichAlias = NULL,
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{type}{Choose "both", "glyco", or "nonGlyco".}

\item{upperLimit}{The upper m/z error limit.}

\item{lowerLimit}{The lower m/z error limit.}

\item{alpha}{The transparency of the dots. Set between 0 and 1.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{silent}{silence printed information.}
}
\value{
A ggplot graph.
}
\description{
A scatter plot showing the retention time versus the m/z error. The three
dotted lines represent the median, and the median plus or minus one standard
deviation. The median and standard deviation calculations will use only points
within the specified upper and lower limits.
}
\examples{
\dontrun{
PlotErrorVersusRt(mydata, type = "both")

PlotErrorVersusRt(mydata, type = "glyco", lowerLimit = -20, upperLimit = 20)
}
}
