% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycanSubcellularLocation.R
\name{PlotGlycanSubcellularLocation}
\alias{PlotGlycanSubcellularLocation}
\title{PlotGlycanSubcellularLocation}
\usage{
PlotGlycanSubcellularLocation(
  input,
  summaryFunction = "count",
  zscoreCutoff = 1.5,
  type = "all",
  pointSize = c(3, 10),
  whichAlias = NULL,
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{summaryFunction}{Either "count" to visualize the number of unique
glycopeptides, or "intensity" to summarise by the log2 intensity of the unique
glycopeptides.}

\item{zscoreCutoff}{An absolute z-score cutoff.}

\item{type}{Either "all" to plot for all glycopeptides, or specify a vector of
glycotypes as in the GlycanType column (e.g., c("Sialylated", "Fucosylated"))}

\item{pointSize}{What is the minimum and maximum point size.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{silent}{silence printed information}
}
\value{
a point plot.
}
\description{
Visualize the count or intensity of glycopeptides per subcellular localization.
}
\examples{
\dontrun{
PlotGlycanSubcellularLocation(mydata, summaryFunction = "intensity", zscoreCutoff = 2,
type = "Fucosylated", pointSize = c(1,10))
}
}
