% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycopeptideCount.R
\name{PlotGlycopeptideCount}
\alias{PlotGlycopeptideCount}
\title{PlotGlycoPeptideCount}
\usage{
PlotGlycopeptideCount(
  input,
  grouping,
  whichAlias = NULL,
  whichPeptide = NA,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{grouping is "technicalReps", "biologicalReps", or "condition"}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{silent}{silence printed information (default = TRUE)}
}
\value{
A graph showing the number of unique glycopeptides
}
\description{
PlotGlycoPeptideCount
}
\examples{
\dontrun{PlotGlycoPSMCount(mydata, grouping = "condition")}
}
