% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycoproteinRank.R
\name{PlotGlycoproteinRank}
\alias{PlotGlycoproteinRank}
\title{PlotGlycoproteinRank}
\usage{
PlotGlycoproteinRank(
  input,
  grouping = c("all", "technicalReps", "biologicalReps", "condition"),
  bins = 10,
  pointSize = 4,
  whichProtein = NULL,
  whichPeptide = NULL,
  exactProteinMatch = TRUE,
  whichAlias = NULL,
  histogramColor = "grey70",
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{grouping}{Grouping is "all" "technicalReps", "biologicalReps",
or "condition".}

\item{bins}{The number of histogram bars.}

\item{pointSize}{The point size in the scatter plot.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{histogramColor}{The color of the histogram bars.}

\item{silent}{silence printed information}
}
\value{
A ggplot object.
}
\description{
Plot the glycoprotein ranks using a scatter plot. The protein intensity is
the summed glycopeptide intensities of that proteins and then log10 transformed.
The histogram shows the total intensity of all glycoproteins per bin.
}
\examples{
\dontrun{
PlotGlycoproteinRank(mydata, grouping = "condition")

PlotGlycoproteinRank(mydata, pointSize = 6,
bins = 20, histogramColor = "#FFFFFF", grouping = "condition")
}
}
