% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycoPSMCount.R
\name{PlotGlycoPSMCount}
\alias{PlotGlycoPSMCount}
\title{PlotGlycoPSMCount}
\usage{
PlotGlycoPSMCount(input, grouping)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{grouping is "technicalReps", "biologicalReps", or "condition"}
}
\value{
A GlycoPSM graph
}
\description{
PlotGlycoPSMCount
}
\examples{
\dontrun{PlotGlycoPSMCount(mydata, grouping = "condition")}
}
