% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycositesVsGlycans.R
\name{PlotGlycositesVsGlycans}
\alias{PlotGlycositesVsGlycans}
\title{PlotGlycositesVsGlycans}
\usage{
PlotGlycositesVsGlycans(
  input,
  whichAlias = NULL,
  pointSize = 3,
  labelGlycositeCutoff = 0,
  labelGlycanCutoff = 0,
  maxOverlaps = 10,
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  plotColor = c("#BAA5CC", "#32006e"),
  alpha = 0.5,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{pointSize}{The size of the points.}

\item{labelGlycositeCutoff}{The minimal number of glycosites found to label
a protein (default = 0)}

\item{labelGlycanCutoff}{The minimal number of glycans found to label a protein
(default = 0)}

\item{maxOverlaps}{The maximum number allowed label overlaps (default = 10)}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{plotColor}{The color of plotted points.}

\item{alpha}{The alpha of the plotted points.}

\item{silent}{silence printed information (default = TRUE)}
}
\value{
A scatter plot comparing the number of glycosites to the number of glycans
for each protein.
}
\description{
Plots the proteins by the number of glycosites on the proteins to the
number of identified glycans identified on that protein.
}
\examples{
\dontrun{
PlotGlycositesVsGlycans(mydata)

PlotGlycositesVsGlycans(mydata, plotColor = "black", alpha = 0.1, maxOverlaps = Inf)}
}
