% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCA.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PlotPCA}
\usage{
PlotPCA(
  input,
  quantType = "normalized",
  dropNoQuant = TRUE,
  minPeptideCoverage = 0.5,
  thresholdMode = "total",
  grouping = "technicalReps",
  type = "glyco",
  returnType = c("PCA", "loadingsPlot", "loadingsData"),
  loadingsPlotLabels = 10,
  label = TRUE,
  whichProtein = NULL,
  whichPeptide = NULL,
  exactProteinMatch = TRUE,
  whichAlias = NULL,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{quantType}{"normalized" for normalized intensity values, and "nonNormalized"
for non-normalized intensity values.}

\item{dropNoQuant}{Remove or keep PSMs without a quantitative value or have a quant
of 0.}

\item{minPeptideCoverage}{Works together with the thresholdMode argument.
If thresholdMode = "total", provide a single numeric value. Values between 0
and 1 are interpreted as proportions (e.g., 0.1 = 10\%), while values >= 1 are
interpreted as absolute counts (e.g., 2 = at least two peptides identified).
If thresholdMode = "group", provide a numeric vector of length two (e.g., c(0.1, 3)).
The first value is interpreted in the same way as for "total" (proportion or
absolute count). The second value specifies the minimum number of groups in
which this threshold must be met.}

\item{thresholdMode}{Character string specifying the filtering mode: "total"
applies a global threshold across all data, while "group" identifies the threshold
within groups.}

\item{grouping}{Grouping is "technicalReps", "biologicalReps", or "condition".}

\item{type}{Either "glyco" or "all".}

\item{returnType}{Determines if the functions returns a PCA plot, a loadings plot,
or the loadings plot data.}

\item{loadingsPlotLabels}{Choose the number of points to label, based on the
magnitude of each point calculated using sqrt(PC1^2 + PC2^2).}

\item{label}{TRUE for showing Alias labels, FALSE for no labels.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{silent}{silence printed information}
}
\value{
A ggplot PCA plot.
}
\description{
PCA plot based on PSM quantification. Missing values are imputed using
kNN using the impute package.
}
\examples{
\dontrun{
     PlotPCA(mydata)

     PlotPCA(mydata, quantType = "normalized", minPeptideCoverage = c(0.4, 2),
     thresholdMode = "group", grouping = "biologicalReps", type = "all")

     PlotPCA(mydata, type = "glyco", returnType = "loadingsPlot",
     loadingsPlotLabels = 20)
}
}
