% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotQuantificationQC.R
\name{PlotQuantificationQC}
\alias{PlotQuantificationQC}
\title{PlotQuantificationQC}
\usage{
PlotQuantificationQC(
  input,
  whichAlias = NULL,
  whichQuantification = "both",
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichQuantification}{choose "both", "corrected", or "raw". Determines
what quantification values to plot.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{silent}{silence printed information (default = FALSE)}
}
\value{
Log2 intensity boxplots
}
\description{
A boxplot showing Log2 intensity values either before or after normalization, or both.
}
\examples{
\dontrun{
PlotQuantificationQC(mydata, whichQuantification = "both")}
}
