% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSiteGlycanComposition.R
\name{PlotSiteGlycanComposition}
\alias{PlotSiteGlycanComposition}
\title{PlotSiteGlycanComposition}
\usage{
PlotSiteGlycanComposition(
  input,
  whichProtein,
  exactProteinMatch = TRUE,
  whichPeptide = NULL,
  whichAlias = NULL,
  domainColors = c("#BAA5CC", "#44AA99", "#a9a9a9", "#22031F", "#433E0E"),
  nodeColors = c("#32006e", "white"),
  horizontalPoints = 50,
  yCorrection = 0.25,
  yNudge = 1,
  boxSpacing = 0.075,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichAlias}{A vector of aliases you want to include, corresponding
to the aliases in your annotation input.}

\item{domainColors}{The colors used for labeling the different domains.}

\item{nodeColors}{The colors used for labeling the glycosites on the protein.}

\item{horizontalPoints}{specify the point size. Larger number is smaller points.}

\item{yCorrection}{move the points closer or further away. smaller values is
less spacing between the points}

\item{yNudge}{nudge the points up}

\item{boxSpacing}{defines the spacing of the box around the points. Larger numbers
mean more spacing}

\item{silent}{TRUE if you want info to be printed, FALSE if not.}
}
\value{
A plot showing glyco site
}
\description{
PlotSiteGlycanComposition
}
\examples{
\dontrun{
PlotSiteGlycanComposition(input = mydata, whichProtein = "P01042",
whichAlias = c("firstSample", "secondSample"))
}
}
