% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSiteGlycanCompositionBar.R
\name{PlotSiteGlycanCompositionBar}
\alias{PlotSiteGlycanCompositionBar}
\title{PlotSiteGlycanCompositionBar}
\usage{
PlotSiteGlycanCompositionBar(
  input,
  grouping = "condition",
  scales = "fill",
  intensity = "raw",
  whichAlias = NULL,
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{grouping}{Grouping is "none", "technicalReps", "biologicalReps", or "condition",
or "all".}

\item{scales}{Controls plot normalization, choose "fill" or "stack".}

\item{intensity}{Raw log2, or log10 transformed intensity values.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{silent}{silence printed information}
}
\value{
A barplot.
}
\description{
Visualize the glycan composition per site.
}
\examples{
\dontrun{
PlotGlycanCompositionBar(mydatam scales = "stack")
}
}
