% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVennDiagram.R
\name{PlotVennDiagram}
\alias{PlotVennDiagram}
\title{PlotVennDiagram}
\usage{
PlotVennDiagram(
  input,
  groups,
  grouping = "condition",
  type = "glyco",
  level = c("peptide", "protein", "glycan"),
  whichAlias = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  whichPeptide = NULL,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{groups}{Supply a vector of two or three groups of choice. This works in
tandem with the grouping argument. If "technicalReps" is selected, grouping
corresponds to the values in the Alias column. For "biologicalReps" this is
the value in the Condition column and the BioReplicate column connected with a
_. For example, Condition_1. For "condition" this is the value in the Condition
column.}

\item{grouping}{Choose "condition", "biologicalReps", or "technicalReps".}

\item{type}{Choose either "glyco" to filter for only glycosylated PSMs, or
select "all".}

\item{level}{Compare either on the levels of "peptide", "protein", or
"glycan".}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{silent}{silence printed information (default = FALSE).}
}
\value{
A ggplot visualization.
}
\description{
Creates a Venn diagram based for findings overlap in peptides, proteins, or
glycans, either for glyco only or all. Only supports comparison between two
or three groups. Two-group comparisons will be scaled proportionally, while
three-way comparisons are not as this is mathemetically only possible in a few
cases.
}
\examples{
\dontrun{
PlotVennDiagram(mydata, groups = c("Treated", "Untreated"), level = "glycan")

PlotVennDiagram(mydata, groups = c("Treated_1", "Treated_2"),
grouping = "biologicalReps", level = "protein")
}
}
