% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMissingness.R
\name{ComputeMissingness}
\alias{ComputeMissingness}
\title{ComputeMissingness}
\usage{
ComputeMissingness(
  input,
  type = "glyco",
  removeAllNA = TRUE,
  whichPeptide = NULL,
  whichAlias = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Your formatted GlycoDiveR data.}

\item{type}{Choose between "glyco" or "all"}

\item{removeAllNA}{Remove peptides with only NA values. This can, for example
happen if you upload abundance values using Perseus or MSstats.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{silent}{silence printed information (default = FALSE)}
}
\value{
The percentage of missing data.
}
\description{
This computes the percentage of missing values in your GlycoDiveR uploaded data.
The computation is done on a (glyco)peptide level. Peptides with a quantitative
value of 0 are considered missing.
}
\examples{
\dontrun{ComputeMissingness(mydata, type = "glyco")}
}
