% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeNumberOfGlycoforms.R
\name{ComputeNumberOfGlycoforms}
\alias{ComputeNumberOfGlycoforms}
\title{ComputeNumberOfGlycoforms}
\usage{
ComputeNumberOfGlycoforms(input, UniprotID, whichAlias = NULL)
}
\arguments{
\item{input}{Formatted data}

\item{UniprotID}{The UniprotID as in the UniprotIDs column}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
The number of potential proteoforms
}
\description{
ComputeNumberOfGlycoforms
}
\examples{
\dontrun{ComputeNumberOfGlycoforms(mydata, "Q9NZQ7")}
}
