% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetComparison.R
\name{GetComparison}
\alias{GetComparison}
\title{GetComparison}
\usage{
GetComparison(input, comparisons, type = "glyco", whichAlias = NULL)
}
\arguments{
\item{input}{formatted data}

\item{comparisons}{the comparisons you would like to include}

\item{type}{Choose "glyco" to use only glycopeptides. Otherwise
choose "all".}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
a comparison dataframe
}
\description{
Computes pvalues, BH-corrected pvalues, and log2 fold changes for the specified
conditions. Comparisons are done on the glycopeptide level using the
max intensity value for duplicate PSMs.
}
\examples{
\dontrun{GetComparison(mydata, list(c("PBS", "Lis")))}
}
