% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlycoDiveR-package.R
\docType{package}
\name{GlycoDiveR-package}
\alias{GlycoDiveR}
\alias{GlycoDiveR-package}
\title{GlycoDiveR: Visualization of glycoproteomic data.}
\description{
GlycoDiveR generates graphics of (quantitative) glycoproteomics datasets directly from search engine output files. It currently supports FragPipe, Byonic, pGlyco, Perseus, and MSstats output files. It converts the data into a generic GlycoDiveR format that incorporates key metadata along with UniProt-derived metadata. The important file is compatible with all GlycoDiveR graphics.
}
\author{
\strong{Maintainer}: Tim S Veth \email{tveth@uw.edu}

}
\keyword{internal}
