% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportByonic.R
\name{ImportByonic}
\alias{ImportByonic}
\title{ImportByonic}
\usage{
ImportByonic(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff,
  glycanScoreCutoff,
  scrape = FALSE,
  removeReverse = TRUE
)
}
\arguments{
\item{path}{The folder path, which will be used to find all "Byonic.xsls" files
in this folder and all subfolders.}

\item{annotation}{The annotation dataframe. Please generate a template using
GetAnnotationTemplate("path", tool = "Byonic")}

\item{fastaPath}{The path to the fasta file that was used.}

\item{peptideScoreCutoff}{The score cutoff from the "Score" column. The cutoff
is the lower limit.}

\item{glycanScoreCutoff}{The score cutoff in the Log Prob column. The score is the
upper limit.}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{removeReverse}{Remove reverse hits as noted with the ">Reverse " tag.}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
This is the importer for Byonic data. It needs to untouched Byonic output, which
is a Excel file with at least a Summary and a Spectra tab. This script will look
for all files in the folder and all subfolders using the name tag "Byonic.xslx".
}
\examples{
\dontrun{ImportByonic(path = "C:/Byonic",
annotation = "C:/annotation.csv",
fastaPath = "C:/fastafile.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 1)}
}
