% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMSFragger.R
\name{ImportMSFragger}
\alias{ImportMSFragger}
\title{ImportMSFragger}
\usage{
ImportMSFragger(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff = 0,
  glycanScoreCutoff = 0.01,
  scrape = TRUE,
  normalization = "median",
  convertFPModCodeToMass = TRUE,
  filterForNoNSequon = FALSE,
  confidenceLevel = FALSE,
  TMT = FALSE,
  cutoffFilter = TRUE,
  dropNoQuant = FALSE,
  minPeptideCoverage = FALSE,
  thresholdMode = c("group", "total")
)
}
\arguments{
\item{path}{Path to search engine output folder}

\item{annotation}{Path to annotation file}

\item{fastaPath}{Path to FASTA file}

\item{peptideScoreCutoff}{Peptide score cutoff, which is the Hyperscore
column in the search engine output file.}

\item{glycanScoreCutoff}{Glycan score cutoff, which is the Glycan Q Value
column in the search engine output file.}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{normalization}{The (glyco)peptide normalization used.
Choose between "median" (default), FP_Normalized, FP_MaxLFQ, or none.
median: performs median normalization.
FP_Normalized: extracts the intensity values in the Intensity columns of the
combined_modified_peptide.tsv files.
FP_MaxLFQ: extracts the intensity values in the MaxLFQ.Intensity columns of the
combined_modified_peptide.tsv files.
none: uses the raw intensity values from the PSM.tsv files.}

\item{convertFPModCodeToMass}{MSFragger uses modification code in peptide
modified sequences. This replaces the code with the mass of the modification.
Keep this enabled when importing MSstats comparison results as MSstats uses
modification masses instead of modification codes. Default = TRUE}

\item{filterForNoNSequon}{Filter for peptides without an N-sequon. Only works
for OPair data (default = FALSE).}

\item{confidenceLevel}{What OPair confidence levels to accept, options are
Level1, Level1b, Level2, Level3 (default = FALSE). Provide like this:
confidenceLevel = c("Level1", "Level1b")}

\item{TMT}{set TRUE or FALSE to select for a TMT experiment.}

\item{cutoffFilter}{Filter for the glycan score, peptide score, N Sequon, and
confidence levels.}

\item{dropNoQuant}{Remove or keep PSMs without a quantitative value or have a quant
of 0.}

\item{minPeptideCoverage}{Works together with the thresholdMode argument.
If thresholdMode = "total", provide a single numeric value. Values between 0
and 1 are interpreted as proportions (e.g., 0.1 = 10\%), while values >= 1 are
interpreted as absolute counts (e.g., 2 = at least two peptides identified).
If thresholdMode = "group", provide a numeric vector of length two (e.g., c(0.1, 3)).
The first value is interpreted in the same way as for "total" (proportion or
absolute count). The second value specifies the minimum number of groups in
which this threshold must be met.}

\item{thresholdMode}{Character string specifying the filtering mode: "total"
applies a global threshold across all data, while "group" identifies the threshold
within groups.}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
This imports search results output by MSFragger/Fragpipe. It will work for
N-glycopeptide searches and O-glycopeptides searches with or without
OPair. During import, this function will look for all the psm.tsv files in the
specified folder and all the subfolders. It will also use the
combined_modified_peptide.tsv if you choose to use FragPipe computed
normalized intensities.
}
\examples{
\dontrun{ImportMSFragger(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 0.05,
scrape = FALSE)}
}
