% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMSFragger.R
\name{ImportMSFragger}
\alias{ImportMSFragger}
\title{ImportMSFragger}
\usage{
ImportMSFragger(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff = 0,
  glycanScoreCutoff = 0.01,
  scrape = FALSE,
  normalization = "median",
  convertFPModCodeToMass = TRUE,
  filterForNoNSequon = FALSE,
  confidenceLevel = FALSE
)
}
\arguments{
\item{path}{Path to search engine output folder}

\item{annotation}{Path to annotation file}

\item{fastaPath}{Path to FASTA file}

\item{peptideScoreCutoff}{Peptide score cutoff}

\item{glycanScoreCutoff}{Glycan score cutoff}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{normalization}{The (glyco)peptide normalization used.
Choose between "median" (default), FP_Normalized, FP_MaxLFQ, or none.
median: performs median normalization.
FP_Normalized: extracts the intensity values in the Intensity columns of the
combined_modified_peptide.tsv files.
FP_MaxLFQ: extracts the intensity values in the MaxLFQ.Intensity columns of the
combined_modified_peptide.tsv files.
none: uses the raw intensity values from the PSM.tsv files.}

\item{convertFPModCodeToMass}{MSFragger uses modification code in peptide
modified sequences. This replaces the code with the mass of the modification.
Keep this enabled when importing MSstats comparison results as MSstats uses
modification masses instead of modification codes. Default = TRUE}

\item{filterForNoNSequon}{Filter for peptides without an N-sequon. Only works
for OPair data (default = FALSE).}

\item{confidenceLevel}{What OPair confidence levels to accept, options are
Level1, Level1b, Level2, Level3 (default = FALSE). Provide like this:
confidenceLevel = c("Level1", "Level1b")}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
ImportMSFragger
}
\examples{
\dontrun{ImportMSFragger(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 0.05,
scrape = FALSE)}
}
