% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMSstatsComparison.R
\name{ImportMSstatsComparison}
\alias{ImportMSstatsComparison}
\title{ImportMSstatsComparison}
\usage{
ImportMSstatsComparison(path, cleanCCarbamidomethylation = TRUE)
}
\arguments{
\item{path}{The filepath to the MSstats output file.}

\item{cleanCCarbamidomethylation}{Removes \[57.0215\] from the MSstats
ModifiedPeptide column (default = TRUE)}
}
\value{
A comparison table with the MSstats comparison results
}
\description{
This imports the MSstats comparison results. It needs the following columns to
work fully: Protein, proteinName, PeptideSequence, Label, log2FC, pvalue, adj.pvalue.
}
\examples{
\dontrun{ImportMSstatsComparison("C:/MSstatsFile.csv")}
}
