% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPerseusAbundances.R
\name{ImportPerseusAbundances}
\alias{ImportPerseusAbundances}
\title{ImportPerseusAbundances}
\usage{
ImportPerseusAbundances(input, PerseusPath, cleanCCarbamidomethylation = TRUE)
}
\arguments{
\item{cleanCCarbamidomethylation}{}
}
\description{
ImportPerseusAbundances
}
