% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportpGlyco.R
\name{ImportpGlyco}
\alias{ImportpGlyco}
\title{ImportpGlyco}
\usage{
ImportpGlyco(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff = 0.01,
  glycanScoreCutoff = 0.01,
  normalization = "median",
  scrape = TRUE,
  cutoffFilter = TRUE,
  dropNoQuant = FALSE,
  minPeptideCoverage = FALSE,
  thresholdMode = c("group", "total")
)
}
\arguments{
\item{path}{The folder path, which will be used to find all
'pGlycoDB-GP-FDR-Pro-Quant-Site.txt' files in this folder and all subfolders.}

\item{annotation}{The annotation dataframe. Please generate a template using
GetAnnotationTemplate("path", tool = "pGlyco")}

\item{fastaPath}{The path to the fasta file that was used.}

\item{peptideScoreCutoff}{The score cutoff for the "PeptideFDR" column. The cutoff
is the upper limit.}

\item{glycanScoreCutoff}{The score cutoff for the "GlycanFDR" column. The score is the
upper limit.}

\item{normalization}{The (glyco)peptide normalization used.
Choose between "none" or "median" (default = "median").}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data.}

\item{cutoffFilter}{Filter for the glycan score, peptide score, N Sequon, and
confidence levels.}

\item{dropNoQuant}{Remove or keep PSMs without a quantitative value or have a quant
of 0.}

\item{minPeptideCoverage}{Works together with the thresholdMode argument.
If thresholdMode = "total", provide a single numeric value. Values between 0
and 1 are interpreted as proportions (e.g., 0.1 = 10\%), while values ≥ 1 are
interpreted as absolute counts (e.g., 2 = at least two peptides identified).
If thresholdMode = "group", provide a numeric vector of length two (e.g., c(0.1, 3)).
The first value is interpreted in the same way as for "total" (proportion or
absolute count). The second value specifies the minimum number of groups in
which this threshold must be met.}

\item{thresholdMode}{Character string specifying the filtering mode: "total"
applies a global threshold across all data, while "group" identifies the threshold
within groups.}
}
\value{
Formatted GlycoDiveR data file.
}
\description{
This is the importer for pGlyco data. It needs to untouched pGlyco output, which
is a .txt file named 'pGlycoDB-GP-FDR-Pro-Quant-Site.txt'. This script will look
for all files in the folder and all subfolders.
}
\examples{
\dontrun{ImportpGlyco(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0.01,
glycanScoreCutoff = 0.01,
scrape = FALSE)}
}
