% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFraggerImporter.R
\name{MSFraggerImporter}
\alias{MSFraggerImporter}
\title{Title}
\usage{
MSFraggerImporter(
  path,
  annotation,
  fastaPath,
  peptideScoreCutoff,
  glycanScoreCutoff,
  scrape = FALSE
)
}
\arguments{
\item{path}{Path to search engine output folder}

\item{annotation}{Path to annotation file}

\item{fastaPath}{Path to FASTA file}

\item{peptideScoreCutoff}{Peptide score cutoff}

\item{glycanScoreCutoff}{Glycan score cutoff}

\item{scrape}{set TRUE/FALSE to use scraping of Uniprot data}
}
\value{
Formatted dataframes
}
\description{
Title
}
\examples{
\dontrun{MSFraggerImporter(path = "Z:/Folder",
annotation = "Z:/Folder/annotation.csv",
fasta = "Z:/fasta.fasta",
peptideScoreCutoff = 0,
glycanScoreCutoff = 0.05,
scrape = FALSE)}
}
