% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotComparisonVolcano.R
\name{PlotComparisonVolcano}
\alias{PlotComparisonVolcano}
\title{PlotComparisonVolcano}
\usage{
PlotComparisonVolcano(
  input,
  whichComparison,
  statisticalCutoff = 0.05,
  log2FCCutoff = 1,
  statistic = "adjpvalue",
  whichLabel = "significant",
  maxOverlaps = 10
)
}
\arguments{
\item{input}{formatted comparison dataframe (e.g. using GlycoDiveR's
ImportMSstatsComparison function)}

\item{whichComparison}{which comparison as in the Label column}

\item{statisticalCutoff}{what statistical cutoff (default = 0.05)}

\item{log2FCCutoff}{what log2 fold-change cutoff (default = 1)}

\item{statistic}{choose "adjpvalue" or "pvalue"}

\item{whichLabel}{what to label in the plot.
choose "significant", "none", or supply a vector with the points to label.
For example c("HPT-149", "SPA3K-N271"). These are internally generated using
the comparison dataframe: paste(Proteins, ModificationID, sep = "-")}

\item{maxOverlaps}{the maximum number of overlapping labels in the volcanoplot}
}
\value{
A volcanoplot
}
\description{
PlotComparisonVolcano
}
\examples{
\dontrun{PlotComparisonVolcano(comparison, whichComparison = "Sal-PBS",
statistic = "pvalue", statisticalCutoff = 0.01)}
}
