% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotComparisonVolcano.R
\name{PlotComparisonVolcano}
\alias{PlotComparisonVolcano}
\title{PlotComparisonVolcano}
\usage{
PlotComparisonVolcano(
  input,
  whichComparison,
  statisticalCutoff = 0.05,
  log2FCCutoff = 1,
  statistic = "adjpvalue",
  whichLabel = "significant",
  plotColors = c("#88CCEE", "#BBBBBB", "#882255"),
  maxOverlaps = 10
)
}
\arguments{
\item{input}{Formatted comparison dataframe (e.g. using GlycoDiveR's
ImportMSstatsComparison function)}

\item{whichComparison}{Which comparison as in the Label column.}

\item{statisticalCutoff}{What statistical cutoff.}

\item{log2FCCutoff}{What log2 fold-change cutoff).}

\item{statistic}{Choose "adjpvalue" or "pvalue".}

\item{whichLabel}{What to label in the plot.
choose "significant", "none", or supply a vector with the points to label.
For example c("HPT-149", "SPA3K-N271"). These are internally generated using
the comparison dataframe: paste(Proteins, ModificationID, sep = "-")}

\item{plotColors}{The colors used in the plot.}

\item{maxOverlaps}{The maximum number of overlapping labels in the volcanoplot}
}
\value{
A volcanoplot
}
\description{
A volcano plot showing Log2 transformed intensity values and either
p values or FDR values.
}
\examples{
\dontrun{
PlotComparisonVolcano(comparison, whichComparison = "Sal-PBS",
statistic = "pvalue", statisticalCutoff = 0.01)}
}
