% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompletenessHeatmap.R
\name{PlotCompletenessHeatmap}
\alias{PlotCompletenessHeatmap}
\title{Plot Completeness Heatmap}
\usage{
PlotCompletenessHeatmap(
  input,
  grouping = "peptide",
  peptideType = "glyco",
  exportDataTo = FALSE,
  whichAlias = NULL,
  whichPeptide = NA
)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{peptide or glyco}

\item{peptideType}{glyco or other}

\item{exportDataTo}{provide path to a folder to export the heatmap as a csv file}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}
}
\value{
Grouped heatmap
}
\description{
Plot Completeness Heatmap
}
\examples{
\dontrun{PlotCompletenessHeatmap(mydata)}
}
