% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompletenessHeatmap.R
\name{PlotCompletenessHeatmap}
\alias{PlotCompletenessHeatmap}
\title{Plot Completeness Heatmap}
\usage{
PlotCompletenessHeatmap(
  input,
  grouping = "peptide",
  peptideType = "glyco",
  exportDataTo = FALSE,
  whichAlias = NULL
)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{peptide or glyco}

\item{peptideType}{glyco or other}

\item{exportDataTo}{provide path to a folder to export the heatmap as a csv file}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
Grouped heatmap
}
\description{
Plot Completeness Heatmap
}
\examples{
\dontrun{PlotCompletenessHeatmap(mydata)}
}
