% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompletenessMatrix.R
\name{PlotCompletenessMatrix}
\alias{PlotCompletenessMatrix}
\title{Plot Completeness Matrix}
\usage{
PlotCompletenessMatrix(
  input,
  peptideType = "glyco",
  plotColors = c("grey90", "#44AA99"),
  collapseTechReps = FALSE,
  whichAlias = NULL,
  whichPeptide = NULL,
  whichProtein = NULL,
  exactProteinMatch = TRUE,
  exportDataTo = FALSE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{peptideType}{Choose "glyco" to only view glycopeptide and "other" to  also
include non-modified peptides.}

\item{plotColors}{The colors used for the matrix The default is c("lightgrey", "darkgreen").}

\item{collapseTechReps}{Do you want to collapse the technical replicates.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting, e.g. whichAlias = c("Alias1", "Alias2").}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{exportDataTo}{If a system folder is provided, a CSV file will be exported
to that folder containing the data presented in the matrix}

\item{silent}{silence printed information (default = FALSE)}
}
\value{
A matrix.
}
\description{
Plot a matrix showing the identified and missed (glyco)peptides. Each row
represents one unique glycopeptide.
}
\examples{
\dontrun{
PlotCompletenessMatrix(mydata)
PlotCompletenessMatrix(mydata, peptideType = "other", silent = FALSE)
}
}
