% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCV.R
\name{PlotCV}
\alias{PlotCV}
\title{PlotCV}
\usage{
PlotCV(
  input,
  CV = c("base", "geometric"),
  horizontalLines = c(0, 20, 50, 100),
  whichProtein = NULL,
  type = c("all", "glyco"),
  returnType = c("plot", "data", "GDdata"),
  returnThreshold = NULL,
  whichPeptide = NULL,
  exactProteinMatch = TRUE,
  whichAlias = NULL,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{CV}{Choose "base" to calculate CV using the non-normalized values, and
"geometric" to calculate the geometric CV using log transformed values.}

\item{horizontalLines}{What horizontal lines to show in the plot.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{type}{Choose "glyco" or "all"}

\item{returnType}{Choose "plot" to return the plot, "data" to return the
CV data as a dataframe (this dataframe can be used to filter any visualization
using the whichPeptide argument), or "GDdata" to return filtered GlycoDiveR
data.}

\item{returnThreshold}{Works in concert with returnType = "data". Defines the
the upper CV threshold.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichAlias}{Provide a vector of Aliases to only select these aliases
for plotting.}

\item{silent}{silence printed information}
}
\value{
A ggplot, dataframe, or GlycoDiveR data
}
\description{
Plot the protein CVs using violing plots. The number of proteins in each group
are labeled. Choose between glyco only, or all, and decide if you want to return
a plot or the underlying data. The returned data can directly be used to filter
other plots using the whichPeptide argument.
}
\examples{
\dontrun{
PlotCV(mydata, type = "glyco", returnType = "plot")

PlotCV(mydata, type = "all", returnType = "GDdata", returnThreshold = 20)
}
}
