% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycanCompositionPie.R
\name{PlotGlycanCompositionPie}
\alias{PlotGlycanCompositionPie}
\title{PlotGlycanCompositionPie}
\usage{
PlotGlycanCompositionPie(
  input,
  grouping,
  scales = "free",
  protein = "all",
  whichAlias = NULL,
  whichPeptide = NA
)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{Grouping is "technicalReps", "biologicalReps", or "condition"}

\item{scales}{Controls plot normalization, choose "fixed" or "free",}

\item{protein}{Use "all" for all proteins, otherwise provide a vector containing
proteins from your data  as in the UniprotIDs column}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}
}
\value{
The glycan composition
}
\description{
PlotGlycanCompositionPie
}
\examples{
\dontrun{PlotGlycanCompositionPie(mydata, grouping = "technicalReps",
protein = c("P10204", "Q92930"))}
}
