% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycansPerSite.R
\name{PlotGlycansPerSite}
\alias{PlotGlycansPerSite}
\title{PlotGlycanPerSite}
\usage{
PlotGlycansPerSite(input, whichAlias = NULL, whichPeptide = NA, silent = FALSE)
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{silent}{silence printed information (default = TRUE)}
}
\value{
The number of glycans per site
}
\description{
This plot shows how many glycans were detected at each site—
that is, how many sites have 1 glycan, 2 glycans, and so on.
}
\examples{
\dontrun{PlotGlycanPerSite(myData)}
}
