% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycansPerSite.R
\name{PlotGlycansPerSite}
\alias{PlotGlycansPerSite}
\title{PlotGlycanPerSite}
\usage{
PlotGlycansPerSite(input, whichAlias = NULL, whichPeptide = NA)
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}
}
\value{
The number of glycans per site
}
\description{
PlotGlycanPerSite
}
\examples{
\dontrun{PlotGlycanPerSite(myData)}
}
