% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycansPerSite.R
\name{PlotGlycansPerSite}
\alias{PlotGlycansPerSite}
\title{PlotGlycanPerSite}
\usage{
PlotGlycansPerSite(input, whichAlias = NULL)
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
The number of glycans per site
}
\description{
PlotGlycanPerSite
}
\examples{
\dontrun{PlotGlycanPerSite(myData)}
}
