% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycansPerSite.R
\name{PlotGlycansPerSite}
\alias{PlotGlycansPerSite}
\title{PlotGlycanPerSite}
\usage{
PlotGlycansPerSite(input)
}
\arguments{
\item{input}{Formatted data}
}
\value{
The number of glycans per site
}
\description{
PlotGlycanPerSite
}
\examples{
PlotGlycanPerSite(myData)
}
