% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycopeptideCount.R
\name{PlotGlycopeptideCount}
\alias{PlotGlycopeptideCount}
\title{PlotGlycoPeptideCount}
\usage{
PlotGlycopeptideCount(input, grouping)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{grouping is "technicalReps", "biologicalReps", or "condition"}
}
\value{
A graph showing the number of unique glycopeptides
}
\description{
PlotGlycoPeptideCount
}
\examples{
\dontrun{PlotGlycoPSMCount(mydata, grouping = "condition")}
}
