% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycoproteinGlycanNetwork.R
\name{PlotGlycoProteinGlycanNetwork}
\alias{PlotGlycoProteinGlycanNetwork}
\title{PlotGlycoproteinGlycanNetwork}
\usage{
PlotGlycoProteinGlycanNetwork(
  input,
  condition = NA,
  type = "N",
  edgeWidth = 1.5,
  verticeSize = c(5, 3),
  whichAlias = NULL,
  highlight = "all"
)
}
\arguments{
\item{input}{Formatted data}

\item{condition}{What condition to plot as listed in the annotation dataframe}

\item{type}{Glycan type "N"}

\item{edgeWidth}{defines the linewidth of the edges}

\item{verticeSize}{defines the size of the glycan and protein
nodes c(3,5) means a glycan node size of 3 and a protein node size of 5}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{highlight}{specify what glycan category to highlight, or use "all" to highlight all}
}
\value{
A glycoprotein to glycan network
}
\description{
PlotGlycoproteinGlycanNetwork
}
\examples{
\dontrun{
PlotGlycoProteinGlycanNetwork(mydata)

PlotGlycoProteinGlycanNetwork(
  mydata,
  condition = "DB_treated",
  edgeWidth = 3,
  verticeSize = c(3, 4)
)
}
}
