% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycoPSMCount.R
\name{PlotGlycoPSMCount}
\alias{PlotGlycoPSMCount}
\title{PlotGlycoPSMCount}
\usage{
PlotGlycoPSMCount(input, grouping, whichAlias = NULL, whichPeptide = NA)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{grouping is "technicalReps", "biologicalReps", or "condition"}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}
}
\value{
A GlycoPSM graph
}
\description{
PlotGlycoPSMCount
}
\examples{
\dontrun{PlotGlycoPSMCount(mydata, grouping = "condition")}
}
