% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycositeIdsPerRun.R
\name{PlotGlycositeIdsPerRun}
\alias{PlotGlycositeIdsPerRun}
\title{PlotGlycositeIdsPerRun}
\usage{
PlotGlycositeIdsPerRun(input, protein)
}
\arguments{
\item{input}{Formatted data}

\item{protein}{Protein listed as in the UniprotIDs column}
}
\value{
A plot displaying the glycosites identified across different runs
}
\description{
PlotGlycositeIdsPerRun
}
\examples{
\dontrun{PlotGlycositeIdsPerRun(mydata, protein = "P17047")}
}
