% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGlycositesVsGlycans.R
\name{PlotGlycositesVsGlycans}
\alias{PlotGlycositesVsGlycans}
\title{PlotGlycositesVsGlycans}
\usage{
PlotGlycositesVsGlycans(
  input,
  whichAlias = NULL,
  labelGlycositeCutoff = 0,
  labelGlycanCutoff = 0,
  maxOverlaps = 10,
  whichPeptide = NA
)
}
\arguments{
\item{input}{The formatted GlycoDiveR data.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{labelGlycositeCutoff}{The minimal number of glycosites found to label
a protein (default = 0)}

\item{labelGlycanCutoff}{The minimal number of glycans found to label a protein
(default = 0)}

\item{maxOverlaps}{The maximum number allowed label overlaps (default = 10)}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}
}
\value{
A scatter plot comparing the number of glycosites to the number of glycans
for each protein.
}
\description{
Plots the proteins by the number of glycosites on the proteins to the
number of identified glycans identified on that protein.
}
\examples{
\dontrun{PlotGlycositesVsGlycans(mydata)}
}
