% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPSMCount.R
\name{PlotPSMCount}
\alias{PlotPSMCount}
\title{PlotPSMCount}
\usage{
PlotPSMCount(inputData, grouping = "technicalReps", whichAlias = NULL)
}
\arguments{
\item{inputData}{Formatted data}

\item{grouping}{grouping is "technicalReps", "biologicalReps", or "condition"}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
the PSM count
}
\description{
PlotPSMCount
}
\examples{
\dontrun{PlotPSMCount(myData)}
}
