% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPSMCount.R
\name{PlotPSMCount}
\alias{PlotPSMCount}
\title{PlotPSMCount}
\usage{
PlotPSMCount(inputData)
}
\arguments{
\item{inputData}{Formatted data}
}
\value{
the PSM count
}
\description{
PlotPSMCount
}
\examples{
PlotPSMCount(myData)
}
