% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPSMsVsTime.R
\name{PlotPSMsVsTime}
\alias{PlotPSMsVsTime}
\title{PlotPSMsVsTime}
\usage{
PlotPSMsVsTime(
  input,
  type = "all",
  bindwidth = 5,
  whichAlias = NULL,
  gradientLength = NA
)
}
\arguments{
\item{input}{your formatted data}

\item{type}{Choose between "allGlyco", "both", "glyco", "combined", "all", or
supply a vector of glycan types, such as c("Multi", "nonGlyco", "Sialyl", "Complex/Hybrid",
"Sialyl+Fucose", "Fucose", "Truncated", "High Mannose", "Paucimannose")}

\item{bindwidth}{the bin width used}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{gradientLength}{Define the LC gradient length. If no length is supplied
the maximum PSM retention time is used}
}
\value{
a lineplot
}
\description{
PlotPSMsVsTime
}
\examples{
\dontrun{PlotPSMsVsTime(mydata, type = "allGlyco", whichAlias = c("NP0-1-r1"),
bindwidth = 1, gradientLength = NA)
}
}
