% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTMQuantification.R
\name{PlotPTMQuantification}
\alias{PlotPTMQuantification}
\title{PlotPTMQuantification}
\usage{
PlotPTMQuantification(input, protein, whichAlias = NULL)
}
\arguments{
\item{input}{Formatted data}

\item{protein}{What protein to quantify the PTM for}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
The PTM quantification for a specific protein
}
\description{
PlotPTMQuantification
}
\examples{
\dontrun{PlotPTMQuantification(mydata, "P07361")}
}
