% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTMQuantification.R
\name{PlotPTMQuantification}
\alias{PlotPTMQuantification}
\title{PlotPTMQuantification}
\usage{
PlotPTMQuantification(
  input,
  protein,
  whichAlias = NULL,
  lineWidth = 2,
  rowFontSize = 12,
  showRowNames = TRUE
)
}
\arguments{
\item{input}{Formatted data}

\item{protein}{What protein to quantify the PTM for}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{lineWidth}{defines the black line around the heatmap cells. Use NA
to remove the line}

\item{rowFontSize}{font size for the row labels}

\item{showRowNames}{set to TRUE or FALSE}
}
\value{
The PTM quantification for a specific protein
}
\description{
PlotPTMQuantification
}
\examples{
\dontrun{PlotPTMQuantification(mydata, "P07361")}
}
