% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTMQuantification.R
\name{PlotPTMQuantification}
\alias{PlotPTMQuantification}
\title{PlotPTMQuantification}
\usage{
PlotPTMQuantification(input, protein)
}
\arguments{
\item{input}{Formatted data}

\item{protein}{What protein to quantify the PTM for}
}
\value{
The PTM quantification for a specific protein
}
\description{
PlotPTMQuantification
}
\examples{
PlotPTMQuantification(mydata, "P07361")
}
