% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTMQuantification.R
\name{PlotPTMQuantification}
\alias{PlotPTMQuantification}
\title{PlotPTMQuantification}
\usage{
PlotPTMQuantification(
  input,
  whichProtein = NULL,
  plotColors = c("#BAA5CC", "#32006e", "white"),
  heatmapColors = c("white", "#88CCEE", "#8877A1", "#882255"),
  exactProteinMatch = TRUE,
  whichAlias = NULL,
  lineWidth = 2,
  rowFontSize = 12,
  showRowNames = TRUE,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data imported through a GlycoDiveR importer.}

\item{whichProtein}{Filter what protein to plot (make sure to only provide one).
These should be the IDs as presented in the UniprotIDs column in your GlycoDiveR data.
This can either be a dataframe with a UniprotIDs column, or a vector with the
UniprotID you want to keep.}

\item{plotColors}{Defines the colors of the barplot.
Default: plotColors = c("#00394c", "#27b56e", "white").}

\item{heatmapColors}{Defines the colors of the heatmap.
Default: heatmapColors = c("white", "#27b56e", "#fdd835", "#d84315").}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{lineWidth}{defines the black line around the heatmap cells. Use NA
to remove the line}

\item{rowFontSize}{font size for the row labels}

\item{showRowNames}{set to TRUE or FALSE}

\item{silent}{silence printed information (default = TRUE)}
}
\value{
The PTM quantification for a specific protein
}
\description{
Visualize the glycan quantification of a single protein using a heatmap.
}
\examples{
\dontrun{
PlotPTMQuantification(mydata, whichProtein = "P07361")

PlotPTMQuantification(mydata, whichProtein = "P07361", rowFontSize = 10,
                      showRowNames = FALSE, lineWidth = 0)}
}
