% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotQuantificationQC.R
\name{PlotQuantificationQC}
\alias{PlotQuantificationQC}
\title{PlotQuantificationQC}
\usage{
PlotQuantificationQC(
  input,
  whichAlias = NULL,
  whichQuantification = "both",
  whichPeptide = NA,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichQuantification}{choose "both", "corrected", or "raw". Determines
what quantification values to plot.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{silent}{silence printed information (default = TRUE)}
}
\value{
Log2 intensity boxplots
}
\description{
PlotQuantificationQC
}
\examples{
\dontrun{PlotQuantificationQC(mydata, whichQuantification = "both")}
}
