% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotQuantificationQC.R
\name{PlotQuantificationQC}
\alias{PlotQuantificationQC}
\title{PlotQuantificationQC}
\usage{
PlotQuantificationQC(input, whichAlias = NULL, whichQuantification = "both")
}
\arguments{
\item{input}{Formatted data}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{whichQuantification}{choose "both", "corrected", or "raw". Determines
what quantification values to plot.}
}
\value{
Log2 intensity boxplots
}
\description{
PlotQuantificationQC
}
\examples{
\dontrun{PlotQuantificationQC(mydata, whichQuantification = "both")}
}
