% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSiteGlycanComposition.R
\name{PlotSiteGlycanComposition}
\alias{PlotSiteGlycanComposition}
\title{PlotSiteGlycanComposition}
\usage{
PlotSiteGlycanComposition(
  input,
  protein,
  whichAlias = NULL,
  horizontalPoints = 50,
  yCorrection = 0.25,
  yNudge = 1,
  boxSpacing = 0.05,
  silent = FALSE
)
}
\arguments{
\item{input}{your formatted data}

\item{protein}{The protein you wish to plot, corresponding to the UniprotIDs
column of your dataframe}

\item{whichAlias}{a vector of aliases you want to include, corresponding
to the aliases in your annotation input}

\item{horizontalPoints}{specify the point size. Larger number is smaller points.}

\item{yCorrection}{move the points closer or further away. smaller values is
less spacing between the points}

\item{yNudge}{nudge the points up}

\item{boxSpacing}{defines the spacing of the box around the points. Larger numbers
mean more spacing}

\item{silent}{TRUE if you want info to be printed, FALSE if not}
}
\value{
A plot showing glyco site
}
\description{
PlotSiteGlycanComposition
}
\examples{
\dontrun{PlotSiteGlycanComposition(input = mydata, protein = "P01042",
whichAlias = c("firstSample", "secondSample"))
}
}
