% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSiteQuantification.R
\name{PlotSiteQuantification}
\alias{PlotSiteQuantification}
\title{Barplot quantification per site}
\usage{
PlotSiteQuantification(
  input,
  whichProtein,
  site,
  whichPeptide = NULL,
  exactProteinMatch = TRUE,
  cutoff = NA,
  whichAlias = NULL,
  silent = FALSE
)
}
\arguments{
\item{input}{The input data as imported through one of the GlycoDiveR importers.}

\item{whichProtein}{Filter what proteins to plot. These are the IDs as presented
in the UniprotIDs column in your GlycoDiveR data. This can either be a dataframe
with a UniprotIDs column, or a vector with the UniprotIDs you want to keep.}

\item{site}{The site as in the ModificationID column.}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{exactProteinMatch}{This is only relevant if you select for proteins using
the whichProtein argument. When set to TRUE (default), your supplied UniprotIDs
must be an exact match to the UniprotIDs in the dataframe. When set to FALSE,
it will select non-exact matches. For example, "P61224" will only match to
"P61224,P62834" when set to FALSE.}

\item{cutoff}{An Intensity column, either in percentage or as an absolute number.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{silent}{TRUE if you want info to be printed, FALSE if not}
}
\value{
A graph.
}
\description{
A barplot showing site glycan quantification.
}
\details{
The mean of technical replicates is calculated and the input data filtered for the user specific glycan and peptide cutoffs.
}
\examples{
\dontrun{
PlotSiteQuantification(mydata, whichProtein = "P04004", site = "N243", cutoff = 1e9)

PlotSiteQuantification(mydata, whichProtein = "P04004", site = "N243", cutoff = "10\%")}
}
