% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSiteQuantification.R
\name{PlotSiteQuantification}
\alias{PlotSiteQuantification}
\title{Barplot quantification per site}
\usage{
PlotSiteQuantification(input, protein, site, cutoff = NA, whichAlias = NULL)
}
\arguments{
\item{input}{The input data as imported through one of the GlycoDiveR importers.}

\item{protein}{The protein defined in the UniprotIDs column.}

\item{site}{The site as in the ModificationID column.}

\item{cutoff}{An Intensity column, either in percentage or as an absolute number.}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}
}
\value{
A graph.
}
\description{
Barplot quantification per site
}
\details{
The mean of technical replicates is calculated and the input data filtered for the user specific glycan and peptide cutoffs.
}
\examples{
\dontrun{PlotSiteQuantification(inputdata, protein = "P04004", site = "N243", cutoff = 1e9)}

\dontrun{PlotSiteQuantification(inputdata, protein = "P04004", site = "N243", cutoff = "10\%")}
}
