% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUpSet.R
\name{PlotUpSet}
\alias{PlotUpSet}
\title{Plot UpSet plots to find identification overlaps}
\usage{
PlotUpSet(
  input,
  grouping = "condition",
  whichAlias = NULL,
  type = "glyco",
  level = "peptide",
  nintersects = 40,
  whichPeptide = NA,
  silent = FALSE
)
}
\arguments{
\item{input}{Formatted data}

\item{grouping}{choose between "condition", "biologicalReps", and
"technicalReps"}

\item{whichAlias}{provide a vector of Aliases to only select these aliases
for plotting}

\item{type}{choose between "glyco" or "all" to include only glycopeptides/proteins
or all peptides/proteins}

\item{level}{choose "peptide" or "protein"}

\item{nintersects}{The maximum number of intersects shown}

\item{whichPeptide}{Filter what peptides to plot. This can either be a dataframe
with a ModifiedPeptide peptide column, or a vector with the ModifiedPeptide sequences
that you want to keep. Inputted data with the comparison importer functions is
directly usable, also after filtering using the FilterComparison function.}

\item{silent}{TRUE if you want info to be printed, FALSE if not.}
}
\value{
An UpSet plot
}
\description{
Plot UpSet plots to find identification overlaps
}
\examples{
\dontrun{
 PlotUpSet(mydata, grouping = "biologicalReps",
 type = "glyco", level = "protein", nintersects = 40)
}
}
